ajuste banco

+++ src/main/resources/db/migration/V2__adicionado_tabela_video.sql
@@ -3,6 +3,6 @@ CREATE TABLE `video` (
   `dat_alt` datetime NOT NULL,
   `dat_ins` datetime NOT NULL,
   `data` datetime NOT NULL,
-  `duracao` bigint(20) NOT NULL,
+  `duracao` decimal(10,2) NOT NULL,
   PRIMARY KEY (`id`) 
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_0900_ai_ci;
